General updates from feedback
Updated the figures for aesthetics and added papers to compare. Updated the conclusion, as well as the discussion and various explanations throughout the paper. All reviewer comments have been considered. Now we are simply reviewing all comments to ensure the previous statement is correct to the best of my ability. Waiting review from Claudia Scarlata.